% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_parameters.R
\name{as_test_parameters}
\alias{as_test_parameters}
\title{Create a test_parameters matrix}
\usage{
as_test_parameters(x)
}
\arguments{
\item{x}{A matrix or a numeric data.frame.}
}
\description{
Converts a matrix or numeric data.frame into a test_parameters object which
is a matrix with a quality of life change for printing.
}
