% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_family.R
\name{build_family}
\alias{build_family}
\title{build_* functions}
\arguments{
\item{test_parameters}{a Matrix containing the items and their parameters as
created by \code{build_test_parameters}.}
}
\description{
The \verb{build_*} functions take as an argument a measurement object and returns
something based on them. These functions can be thought of as being,
partially, the  counterpart of the \verb{get_*} functions.

\code{\link[=build_qmatrix]{build_qmatrix()}} and \code{\link[=build_test_parameters]{build_test_parameters()}} work as a pair. The former
takes a \link{test_parameters} matrix and returns a Q-matrix, and the latter
does the opposite.
\itemize{
\item \code{\link[=build_profile_likelihood]{build_profile_likelihood()}} Given a \code{response} returns the likelihood of
each possible profile.
\item \code{\link[=build_response_likelihood]{build_response_likelihood()}} Given a \code{profile} returns the likelihood of
each possible response pattern.
}
}
\keyword{internal}
