% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_response_likelihood.R
\name{build_marginal_response_likelihood}
\alias{build_marginal_response_likelihood}
\title{Builds the marginal response likelihood of the given response patterns}
\usage{
build_marginal_response_likelihood(
  test_parameters,
  responses,
  priors,
  complete = TRUE
)
}
\arguments{
\item{test_parameters}{a Matrix containing the items and their parameters as
created by \code{build_test_parameters}.}

\item{responses}{numeric vector or matrix where each row corresponds to a
response pattern.}

\item{priors}{a vector indicating the priors for each profile. Use
\code{\link[=create_patterns]{create_patterns()}} to see the order in which you should provide the priors.}

\item{complete}{Should the result include the response patterns?
default \code{TRUE}.}
}
\value{
a matrix
}
\description{
This function returns the likelihood of a given response marginal on all
possible profiles.
}
\details{
\deqn{
   P(\bold{x}) =
     \sum^L_{l=1} P(\bold{x} \mid \bold{\alpha}_l) P(\bold{\alpha}_l)
   }

If more than one response pattern is passed, then the result is a matrix
where each row contains the response pattern and it's respective
likelihood.
}
