% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_profile_likelihood.R
\name{build_profile_likelihood}
\alias{build_profile_likelihood}
\title{Build profiles likelihood and posterior probabilities for a given response.}
\usage{
build_profile_likelihood(
  test_parameters,
  profiles,
  response,
  profile_priors,
  priors,
  complete = TRUE
)
}
\arguments{
\item{test_parameters}{a Matrix containing the items and their parameters as
created by \code{build_test_parameters}.}

\item{profiles}{Vector or matrix with the profiles.}

\item{response}{numeric vector corresponding to the response pattern.}

\item{priors}{profiles prior distribution.}

\item{profile_prior}{the priors of the passed profiles}
}
\value{
A matrix with the profiles, the priors, the likelihood and the
posterior
}
\description{
This function returns the likelihood of a profile given a response:
\eqn{P(\bold{\alpha_l}\mid\bold{x})}.
}
