% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_qmatrix.R
\name{build_qmatrix}
\alias{build_qmatrix}
\title{Build a Q-Matrix from a a matrix of test parameters}
\usage{
build_qmatrix(test_parameters)
}
\arguments{
\item{test_parameters}{a Matrix containing the items and their parameters as
created by \code{build_test_parameters}.}

\item{attributes_names}{vector of names for the attributes. Defaults
to "Attr#"}

\item{item_names}{vector of names for the items. Defaults to "Item#"}
}
\value{
a Q-matrix
}
\description{
Build a Q-Matrix from a a matrix of test parameters
}
