% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_test_parameters.R
\name{build_test_parameters}
\alias{build_test_parameters}
\title{Create a matrix of items and their parameters}
\usage{
build_test_parameters(qmat, ...)
}
\arguments{
\item{qmat}{Q-matrix}

\item{...}{values to use in the different factor levels. i.e. the first
number is assumed to be the intercept, then main effects, then two-way
interaction and so on.}
}
\value{
A matrix of items and their parameters.
}
\description{
This matrix already encodes the Q-matrix. It can be useful to save the
results for future simulations or further wrangling as a csv.
}
