% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_matrix.R
\name{change_matrix}
\alias{change_matrix}
\title{Change Matrix}
\usage{
change_matrix(mat, fun, at = NULL, direction = "row")
}
\arguments{
\item{mat}{Matrix to change}

\item{fun}{function to apply}

\item{at}{vector of indices, logic vector or function to determine the
rows/columns to be changed.}

\item{direction}{should the change be done for the rows or columns.}
}
\value{
A matrix of the same dimension than \code{mat}.
}
\description{
Apply a function to the rows/columns of a matrix. This function is just a
wrapper around the \code{apply} function with some quality of life changes. It
is designed to be used mainly to modify Q-matrices and test matrices using
the defined utility functions provided in the package.
}
