% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_test.R
\name{change_test}
\alias{change_test}
\title{Change a test parameters}
\usage{
change_test(
  test_parameters,
  level = NULL,
  attribute = NULL,
  items = NULL,
  values = NULL,
  fun = NULL
)
}
\arguments{
\item{test_parameters}{A matrix}

\item{level}{The level of parameters to change. 0 corresponds to the
intercept, 1 to the main effects, 2 to the two-way interaction and so on.}

\item{attribute}{The attribute that must be present in the parameter.}

\item{items}{Indices of items to be changed.}

\item{values}{Values to replace the selected values. If there are more
values than numbers to replace, then it will just use}

\item{fun}{A function that takes all values to be changed and returns a
vector of numbers of the same length.}
}
\value{
A matrix corresponding to the test parameters with the changes.
}
\description{
Change a test parameters
}
