% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmatrix.R
\name{check_qmatrix}
\alias{check_qmatrix}
\title{Q-Matrix composition}
\usage{
check_qmatrix(qmat)
}
\arguments{
\item{qmat}{Qmatrix to analyze}
}
\value{
Summary like text
}
\description{
Check the composition of a Q-Matrix
}
