% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_test_parameters.R
\name{check_test_parameters}
\alias{check_test_parameters}
\title{Check if the test parameters are compatible with a Q-matrix}
\usage{
check_test_parameters(
  test,
  qmatrix = NULL,
  action = c("error", "warning", "message", "quiet")
)
}
\arguments{
\item{test}{item parameters matrix}

\item{qmatrix}{Optional. If no qmatrix is passed, then it checks that it is
compatible for some qmatrix}
}
\value{
boolean
}
\description{
Check if the test parameters are compatible with a Q-matrix
}
