% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{construct_parameters_labels}
\alias{construct_parameters_labels}
\title{Construct the labels for all parameters}
\usage{
construct_parameters_labels(attr_labels, null_profile = "0", collapse = "-")
}
\arguments{
\item{attr_labels}{vector with the labels of each attribute}

\item{null_profile}{character for the null profile (all zeroes)}

\item{collapse}{character to use when collapsing the attributes.}
}
\value{
a vector with parameters
}
\description{
This is an utility function that helps generating all the labels for
possible parameters.
}
