% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_item.R
\name{create_item}
\alias{create_item}
\title{Create item parameters from probabilities of right response}
\usage{
create_item(..., type = "LCDM", num_attrs = NULL)
}
\arguments{
\item{...}{probabilities of right answer by attribute profile.}

\item{type}{type of item (LCDM, DINA, DINO, RRUM).}

\item{num_attrs}{Number of attributes in the item. Not needed for LCDM.}
}
\value{
an item parameters in their standard format
}
\description{
Create item parameters from probabilities of right response
}
\details{
This function returns a vector of item parameters from a vector
of probabilities of having right response for a specific attribute vector.
Notice: This function doesn't check for  possible \emph{real} values
(e.g. main effects could be negative).
}
\examples{
lcdm_item <- create_item(.2, .4, .6, .8)
rrum_item <- create_item(.2, .4, .6, type = "RRUM")
dino_item <- create_item(.2, .8, type = "DINO", num_attrs = 2)
}
