% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_items_from_qmat.R
\name{create_items_from_qmat}
\alias{create_items_from_qmat}
\title{Creates a matrix of items-parameters given a qmatrix.}
\usage{
create_items_from_qmat(qmat, ...)
}
\arguments{
\item{qmat}{Q-matrix}

\item{...}{Values for the main effects, two interaction and so on. The values are all assumed equal. Do not need to pass interaction effects that don't exist in the Q-matrix.}
}
\value{
A matrix where the rows are the items and the columns all the possible parameters.
}
\description{
This is an utility function to quickly create a matrix of parameters that is consistent with a given qmatrix. If no extra parameters are passed, then it returns a mask for the parameters where each number represents a factors order (e.g. 1 indicatens main effect, 2 indicates two-way effect, and so on.).
}
