% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmatrix.R
\name{create_qmatrix}
\alias{create_qmatrix}
\title{Create Q-Matrix}
\usage{
create_qmatrix(total_attrs = 3L, item_attrs = 1L, repeat_items = 1L)
}
\arguments{
\item{total_attrs}{attributes in the Q-matrix}

\item{item_attrs}{How many attributes should the items have}

\item{repeat_items}{Items per unique combination}
}
\value{
a Q-matrix
}
\description{
Create Q-Matrix
}
\examples{
# This returns a simple Q-matrix of 3 attributes
create_qmatrix(3)
# This returns a simple Q-matrix of 3 attributes with 2 items per attribute
create_qmatrix(3,2)
# This makes the items be associated with two attributes
create_qmatrix(3,item_attrs=2)
}
