% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_response_likelihood.R
\name{create_response_likelihood}
\alias{create_response_likelihood}
\title{Returns the likelihood and probability of mastery of the response patterns.}
\usage{
create_response_likelihood(
  items,
  response_patterns = NULL,
  prior = NULL,
  EAP_treshold = 0.5
)
}
\arguments{
\item{items}{a matrix of probabilities for non-masters and masters. Each row is an item.}

\item{response_patterns}{A binary matrix where each row is one response pattern. If none is passed, then all possible patterns are generated}

\item{prior}{The prior for non-mastery and mastery ratio.}

\item{EAP_treshold}{Posterior threshold to be considered a master. By default is .5}
}
\value{
a matrix where the first columns contain the response patterns, the likelihood of seeing
that pattern given a certain level of mastery. The probability of being a (non)master and the EAP (final classification)
}
\description{
Returns the likelihood and probability of mastery of the response patterns.
}
