% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test.R
\name{create_test}
\alias{create_test}
\title{Create a matrix of items}
\usage{
create_test(...)
}
\arguments{
\item{...}{integer vectors with the probabilities}
}
\value{
a matrix of parameters
}
\description{
@description
An utility function that calls the \code{create_item} function and returns a matrix
padded with NAs that can be used directly with the responses generation.
It assumes all the items are LCDM.
}
