% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.item_influence.R
\docType{data}
\name{data.item_influence}
\alias{data.item_influence}
\title{Item Influence for DCM.}
\format{
An object of class \code{list} of length 3.
}
\usage{
data.item_influence
}
\description{
Q-matrix and simulated response data based in the Jurich & Madison (2023)
simulation.
}
\references{
Jurich, D. P., & Madison, M. J. (2023). Measuring Item Influence
for Diagnostic Classification Models. Educational Assessment, 28(4),
229–245. https://doi.org/10.1080/10627197.2023.2244411
}
\keyword{datasets}
