% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ids.R
\name{.generate_ids}
\alias{.generate_ids}
\title{Internal - Generate random simulation ids}
\usage{
.generate_ids(n = 1L)
}
\arguments{
\item{n}{number of ids to generate}
}
\value{
a vector of random ids to be used in simulations
}
\description{
Internal - Generate random simulation ids
}
\keyword{internal}
