% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sample.R
\name{.generate_sample}
\alias{.generate_sample}
\title{Internal - Generate sample of respondents
Note: Not encapsulated for randomness.}
\usage{
.generate_sample(
  sample_size,
  total_attrs,
  base_rate = 0.5,
  attr_corr = 0.5,
  attributes.names = NULL,
  responses.names = NULL
)
}
\arguments{
\item{sample_size}{How many respondents}

\item{total_attrs}{How many attributes are in the sample}

\item{base_rate}{Base rate per attribute (or one if it's the same for all)}

\item{attr_corr}{Correlation of attributes (or one if it's the same for all)}

\item{attributes.names}{vector of names for the attributes. Defaults to "Attr#"}

\item{responses.names}{vector of names for the responses. Defaults to "#"}
}
\value{
A matrix of respondents and attributes.
}
\description{
Internal - Generate sample of respondents
Note: Not encapsulated for randomness.
}
