% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_attr_mask_from_profile}
\alias{.get_attr_mask_from_profile}
\title{Get attribute matrix mask from profiles}
\usage{
.get_attr_mask_from_profile(profiles)
}
\arguments{
\item{profiles}{a matrix of the profiles for which we want the masks}
}
\value{
A matrix
}
\description{
Returns a matrix where each row contains the parameters for each given
profile
}
