% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_matching_rows_index}
\alias{.get_matching_rows_index}
\title{Utility function to filter rows in matrices}
\usage{
.get_matching_rows_index(A, B)
}
\arguments{
\item{A}{Complete matrix to filter.}

\item{B}{subset of rows to match in A. Can also be a single vector.}
}
\value{
a vector of indices in A that match B.
}
\description{
Utility function to filter rows in matrices
}
