% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{.get_parameter_item_matrix}
\alias{.get_parameter_item_matrix}
\title{Qmatrix to parameters-items matrix}
\usage{
.get_parameter_item_matrix(qmat, items)
}
\arguments{
\item{qmat}{Qmatrix}

\item{items}{matrix of items-parameters}
}
\value{
parameter-item matrix
}
\description{
Qmatrix to parameters-items matrix
}
