% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{.get_prob_matrix}
\alias{.get_prob_matrix}
\title{Respondents-items matrix}
\usage{
.get_prob_matrix(qmat, respondents, items)
}
\arguments{
\item{qmat}{Q-matrix}

\item{respondents}{Binary matrix of respondents}

\item{items}{dataframe of items}
}
\value{
a matrix of responses
}
\description{
Respondents-items matrix
}
