% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_matrix.R
\name{.is_binary_row}
\alias{.is_binary_row}
\title{Check if vector is binary}
\usage{
.is_binary_row(row)
}
\arguments{
\item{row}{a numeric vector}
}
\value{
boolean vector
}
\description{
Check if vector is binary
}
