% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{.parameter_attr_matrix}
\alias{.parameter_attr_matrix}
\title{items parameter - attribute matrix}
\usage{
.parameter_attr_matrix(num_att)
}
\arguments{
\item{num_att}{Number of attributes}
}
\value{
a matrix
}
\description{
items parameter - attribute matrix
}
