% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{.parameter_mask}
\alias{.parameter_mask}
\title{Parameter mask}
\usage{
.parameter_mask(attr_vector, num_attrs = NULL, add_intercept = T)
}
\arguments{
\item{attr_vector}{binary vector showing presence of attribute}

\item{num_attrs}{number of attributes if longer than the passed vector}

\item{add_intercept}{if the intercept is added at the start of the result}
}
\value{
A binary vector with all the parameter that the item should include
}
\description{
Parameter mask
}
