% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_simulation_seed.R
\name{.seed_from_id}
\alias{.seed_from_id}
\title{Create simulation seed from ID}
\usage{
.seed_from_id(id = NULL)
}
\arguments{
\item{id}{An arbitrary R object which will then be hashed and converted to an int.}
}
\value{
The function returns a seed to be used.
}
\description{
This function is meant to be used inside the generation functions for a
finer replication of simulations. It works by taking an ID (any kind of R object),
hashing it, converting to an integer and setting that as a seed.
}
