% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmatrix.R
\name{extend_qmatrix}
\alias{extend_qmatrix}
\title{Extend a Q-Matrix}
\usage{
extend_qmatrix(.qmat, item_attrs = 1L, repeat_items = 1L)
}
\arguments{
\item{.qmat}{Qmatrix to be extended}

\item{item_attrs}{How many attributes should the items have}

\item{repeat_items}{Items per unique combination}
}
\value{
An extended version of the passed Q-Matrix
}
\description{
This function takes a Q-Matrix and adds items with certain
amount of attributes.
}
\examples{
# This creates three more items with two attributes each
create_qmatrix(3) |>
  extend_qmatrix(1, 2)
}
