% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lcdm.R
\name{fit_lcdm}
\alias{fit_lcdm}
\title{Utility function to fits a LCDM model using the CDM package with some
convenient defaults.}
\usage{
fit_lcdm(
  responses,
  qmatrix,
  monotonicity = TRUE,
  rule = "GDINA",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{responses}{examinees responses}

\item{qmatrix}{qmatrix.}

\item{monotonicity}{Should the monotonicity constraint be forced?}

\item{rule}{model to use in the gdina. defaults to GDINA}

\item{verbose}{should it print progress? defaults to FALSE.}

\item{...}{extra parameters to pass to the gdina call.}
}
\value{
an object gdina.
}
\description{
Utility function to fits a LCDM model using the CDM package with some
convenient defaults.
}
