% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_models}
\alias{fit_models}
\title{Fit many regression models based on a vector of formulas}
\usage{
fit_models(
  formulas,
  fitter = lm,
  ...,
  output = FALSE,
  print = TRUE,
  texreg = list()
)
}
\arguments{
\item{formulas}{A vector of \code{formulas}.}

\item{fitter}{fitting function or vector of fitting functions (default: \code{lm}).}

\item{...}{other arguments passed to fitting function.}

\item{output}{Return estimated models in a list?}

\item{print}{Print a regression table?}

\item{texreg}{A named list with arguments passed to \code{screenreg} if \code{print=TRUE}.}
}
\value{
either \code{NULL} or the list of fitted model objects.
}
\description{
Fit many regression models based on a vector of formulas
}
\details{
This simple function estimates a series of regression models based on a vector of formulas.
The argument \code{fitter} allows to call any formula-based fitting function in R (e.g., lm, glm).
The function provides options to summarize the regressions in a regression table and/or to store
the fit objects in a named list which can passed on.
}
\examples{

data(mtcars)

f_lm <- c(
	mpg ~ vs, 
	hp ~ vs, 
	disp ~ vs * am)

fit_models(f_lm, fitter=lm, data=mtcars, 
	output=FALSE, print=TRUE)

f_glm <- c(
	vs ~ mpg, 
	am ~ mpg)

mods <- fit_models(f_glm, fitter=glm, data=mtcars, 
	family=binomial('logit'), 
	output=TRUE, print=TRUE, 
	texreg=list(stars=NULL))

lapply(mods, nobs)



}
\seealso{
\code{\link[texreg]{screenreg}}
}
