% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_examinees.R
\name{generate_examinees}
\alias{generate_examinees}
\title{Generate sample of examinees}
\usage{
generate_examinees(
  sample_size,
  total_attrs,
  base_rate = 0.5,
  attr_corr = 0,
  strict = TRUE,
  tolerance = NULL,
  attributes_names = NULL,
  responses_names = NULL
)
}
\arguments{
\item{sample_size}{How many examinees to generate}

\item{total_attrs}{How many attributes are in the sample}

\item{base_rate}{Base rate per attribute (or one if it's the same for all)}

\item{attr_corr}{Correlation of attributes (or one if it's the same for all)}

\item{strict}{Boolean. If true, the function tries to generate a
binary matrix with the given correlation. If it is not possible, then it
raises and error. If false it is considered to be the correlation matrix of
the underlying normal distributions.}

\item{tolerance}{When strict_binary_corr=FALSE, the actual correlation values
can be far from the expected values. This gives a tolerance to avoid
unexpected results.}

\item{attributes_names}{vector of names for the attributes. Defaults to
"Attr#"}

\item{responses_names}{vector of names for the responses. Defaults to "ID#"}
}
\value{
A matrix of respondents and attributes.
}
\description{
Generate sample of examinees
}
