% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_examinees.R
\name{generate_examinees_MVN}
\alias{generate_examinees_MVN}
\title{Generate examinees using a Multivariate Normal distribution}
\usage{
generate_examinees_MVN(sample_size, total_attrs, base_rate, attr_corr)
}
\arguments{
\item{base_rate}{Ratio of examinees that master each attribute. If only one
number is supplied then it's assumed is the same for all.}

\item{attr_corr}{Correlation of attributes. The order of the correlations is
assumed to be 1-2, 1-3, ...., 2-3, 2-4, and so on.}
}
\description{
Generate examinees using a Multivariate Normal distribution
}
\keyword{internal}
