% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_id.R
\name{generate_id}
\alias{generate_id}
\title{Generate random simulation ids}
\usage{
generate_id(n = 1L)
}
\arguments{
\item{n}{number of ids to generate}
}
\value{
a vector of random ids to be used in simulations
}
\description{
Generate random simulation ids
}
