% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ids.R
\name{generate_ids}
\alias{generate_ids}
\title{Generate random simulation ids}
\usage{
generate_ids(n = 1L, id = NULL)
}
\arguments{
\item{n}{number of ids to generate}

\item{id}{simulation id, to be used as a seed within this function.}
}
\value{
a vector of random ids to be used in simulations
}
\description{
Generate random simulation ids
}
