% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_responses.R
\name{generate_responses}
\alias{generate_responses}
\title{Generate responses to a test}
\usage{
generate_responses(
  examinees,
  test,
  get_probs = FALSE,
  method = c("DCM", "IRT", "MIRT", "UDF"),
  FUN = NULL,
  ...
)
}
\arguments{
\item{examinees}{Binary matrix of respondents}

\item{test}{matrix of items parameters}

\item{get_probs}{if you want the probability of correct response matrix.}

\item{method}{What method to use for the generation of responses.}

\item{FUN}{If \code{method="UDF"}, then the function will use a user provided
function for response generation.}
}
\value{
a matrix of responses in the form 0/1 or the real probabilities.
}
\description{
Generate responses to a test
}
