% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{generate_sample}
\alias{generate_sample}
\title{Generate sample of respondents (categorical data)}
\usage{
generate_sample(sample_size, total_attrs, base_rate = 0.5, attr_corr = 0.5)
}
\arguments{
\item{sample_size}{How many respondents}

\item{total_attrs}{How many attributes are in the sample}

\item{base_rate}{Base rate per attribute (or one if it's the same for all)}

\item{attr_corr}{Correlation of attributes (or one if it's the same for all)}
}
\value{
A matrix of respondents and attributes.
}
\description{
Generate sample of respondents (categorical data)
}
