% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_examinees.R, R/get_family.R,
%   R/get_profiles_ratio.R, R/get_test_parameters.R
\name{get_examinees}
\alias{get_examinees}
\alias{get_family}
\alias{get_profiles_ratio}
\alias{get_test_parameters}
\title{get_* Functions}
\usage{
get_examinees(model, method = c("map", "mle"))

get_profiles_ratio(model)

get_test_parameters(model)
}
\arguments{
\item{model}{A fitted GDINA model.}

\item{method}{The method used in classification.}
}
\value{
A matrix with the MAP or MLE classifications of examinees.

A matrix with all the profiles in the model and their estimated
ratio

a \link{test_parameters} object from the parameters estimated by the
model, which is just an extension of the base \link[base:matrix]{base::matrix} class.
}
\description{
The \verb{get_*} functions take as a first argument a fitted GDINA model and
return a matrix to simplify piping operations.
\itemize{
\item \code{\link[=get_test_parameters]{get_test_parameters()}} returns a \link{test_parameters} object from the
parameters estimated by the model.
\item \code{\link[=get_examinees]{get_examinees()}} returns a numeric matrix of the students classification
by attribute.
}
}
\concept{examinees functions}
\concept{profiles functions}
\concept{test parameters functions}
\keyword{internal}
