% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ideal_patterns.R
\name{get_ideal_patterns}
\alias{get_ideal_patterns}
\title{Get the matrix of the possible ideal responses patterns from a qmatrix}
\usage{
get_ideal_patterns(qmat, rule)
}
\arguments{
\item{qmat}{Q-matrix from which to generate the patterns}

\item{rule}{DINA (only students with all relevant attributes should have the item correct) or DINO (students with at least one relevant attribute should have the item correct)}
}
\value{
a matrix of responses patterns and items
}
\description{
Get the matrix of the possible ideal responses patterns from a qmatrix
}
