% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_blueprint.R
\name{get_information}
\alias{get_information}
\title{Get attribute information}
\usage{
get_information(model, responses)
}
\arguments{
\item{model}{a fitted gdina model}

\item{responses}{the response data}
}
\value{
a list with the info and the info proportion by attribute
}
\description{
Get attribute information
}
