% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_item_parameters.R
\name{get_item_parameters}
\alias{get_item_parameters}
\title{Returns a matrix item-parameters from model}
\usage{
get_item_parameters(model, print_table = TRUE, digits = 2)
}
\arguments{
\item{model}{a GDINA object. It has to be logit link function.}

\item{print_table}{boolean. If true it prints a pretty table. It returns a data frame silently.}

\item{digits}{rounding digits when printing. The return is not approximated.}
}
\value{
matrix of item-parameters
}
\description{
This extracts a model's item parameters as a dataframe. If \code{print_table} is true, then it also prints a pretty version of the table.
}
