% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_parameters.R
\name{get_test_parameters}
\alias{get_test_parameters}
\title{Returns a matrix item-parameters from model}
\usage{
get_test_parameters(model, pretty_print = TRUE, digits = 2)
}
\arguments{
\item{model}{a GDINA object. It has to be logit link function.}

\item{pretty_print}{boolean. If true it prints a pretty table. It returns a data frame silently.}

\item{digits}{rounding digits when printing. The return is not approximated.}
}
\value{
matrix of item-parameters
}
\description{
This extracts a model's item parameters as a matrix If \code{pretty_print} is
true, then it also prints a pretty version of the table.
}
