% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{item_influence}
\alias{item_influence}
\title{Item Influence for DCM}
\format{
A numeric matrix with 100 rows and 5 columns.
}
\source{
\url{https://doi.org/10.1080/10627197.2023.2244411}
}
\usage{
item_influence
}
\description{
Q-matrix and simulated response data based on Jurich & Madison (2023).
}
\keyword{datasets}
