% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{label_matrix}
\alias{label_matrix}
\title{Utility to label matrices}
\usage{
label_matrix(
  mat,
  col_prefix = "A",
  row_prefix = "V",
  col_labels = NULL,
  row_labels = NULL
)
}
\arguments{
\item{mat}{Matrix to label}

\item{col_prefix, row_prefix}{prefix to use to create default labels}

\item{col_labels, row_labels}{vector with the row/column labels}
}
\value{
The original matrix with labels
}
\description{
Utility to label matrices
}
\keyword{internal}
