% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_parameters.R
\name{pretty_print}
\alias{pretty_print}
\title{Prints a matrix to the console in a prettier format}
\usage{
pretty_print(matrix, digits = 3, prefix = NULL)
}
\arguments{
\item{matrix}{matrix to print}

\item{digits}{number of rounding digits}

\item{prefix}{prefix for the columns.By default, uses the Greek letter
lambda in interactive sessions.}
}
\value{
the same passed object
}
\description{
This function prints a pretty table and returns the original object
so it can be used inside a pipeline.
}
