% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_responses.R
\name{probability_functions}
\alias{probability_functions}
\alias{DCM_prob_fn}
\alias{IRT_prob_fn}
\alias{MIRT_prob_fn}
\title{Probability Functions for Response Generation}
\usage{
DCM_prob_fn(examinees, test)

IRT_prob_fn(examinees, test)

MIRT_prob_fn(examinees, test)
}
\arguments{
\item{examinees}{A matrix of respondent traits or abilities.}

\item{test}{A matrix of item parameters.}
}
\value{
A matrix of probabilities where rows correspond to examinees
and columns correspond to items.
}
\description{
These internal functions calculate the item response probability matrices
for different psychometric frameworks.
}
\keyword{internal}
