% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_test_parameters.R
\name{read_test_parameters}
\alias{read_test_parameters}
\title{Read test parameters file}
\usage{
read_test_parameters(path, monotonicity_check = TRUE, ...)
}
\arguments{
\item{path}{Path of the .csv file with the test parameters.}

\item{monotonicity_check}{Boolean. Check the monotonicity of the test using
\code{\link[=check_monotonicity]{check_monotonicity()}}.}

\item{...}{Extra arguments to be passed to \code{\link[=read.csv]{read.csv()}}.}
}
\value{
A matrix of the test parameters.
}
\description{
Read test parameters file
}
