% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_simulation_seed.R
\name{set_simulation_seed}
\alias{set_simulation_seed}
\title{Set simulation seed}
\usage{
set_simulation_seed(id = NULL)
}
\arguments{
\item{id}{An arbitrary R object which will then be hashed and converted to an int.}
}
\value{
The function returns the seed that was setted.
}
\description{
This function is meant to be used inside the generation functions for a
finer replication of simulations. It works by taking an ID (any kind of R object),
hashing it, converting to an integer and setting that as a seed.
}
\examples{
# We choose an id an the seed will be set
set_simulation_seed(id = "my-simulation-id")
a <- runif(10)
set_simulation_seed(id = "my-simulation-id")
b <- runif(10)
a == b # These values will be equal
}
