% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_parameters_example}
\alias{test_parameters_example}
\title{LCDM test parameters example}
\format{
A data frame with 50 observations.
}
\source{
Created as an example
}
\usage{
test_parameters_example
}
\description{
Example of a test parameters matrix.
}
\keyword{datasets}
