% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_parameters.R
\name{tidy_parameters}
\alias{tidy_parameters}
\title{Convert the parameters of a model and/or test into a tibble for easier analysis}
\usage{
tidy_parameters(model = NULL, qmat = NULL, test = NULL)
}
\arguments{
\item{model}{fitted gdina model}

\item{qmat}{Q-matrix}

\item{test}{test matrix (obtained from \link{create_test})}
}
\value{
a tibble with the estimation and/or real values of the items.
}
\description{
Convert the parameters of a model and/or test into a tibble for easier analysis
}
