% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_profile_likelihood.R
\name{build_profile_likelihood}
\alias{build_profile_likelihood}
\title{Build profiles likelihood and posterior probabilities for a given response.}
\usage{
build_profile_likelihood(
  test_parameters,
  response,
  num_attributes = log(ncol(test_parameters), 2),
  priors = NULL
)
}
\arguments{
\item{test_parameters}{a Matrix containing the items and their parameters as
created by \code{build_test_parameters}. If the test is extracted from a model
using \code{get_test_parameters} it should be extracted using the parameter
\code{complete=TRUE}.}

\item{response}{numeric vector corresponding to the response pattern.}

\item{num_attributes}{The number of attributes in the test. It will be
inferred from the test.}

\item{priors}{profiles prior distribution. If NULL it will be assumed to be
uniformly distributed.}
}
\value{
A matrix with the profiles, the priors, the likelihood and the
posterior
}
\description{
Build profiles likelihood and posterior probabilities for a given response.
}
