% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_response_likelihood.R
\name{build_response_likelihood}
\alias{build_response_likelihood}
\title{Create the likelihood matrix of all possible response given a profile.}
\usage{
build_response_likelihood(test_parameters, profile)
}
\arguments{
\item{test_parameters}{a Matrix containing the items and their parameters as
created by \code{build_test_parameters}. If the test is extracted from a model
using \code{get_test_parameters} it should be extracted using the parameter
\code{complete=TRUE}.}

\item{profile}{numeric vector corresponding to the assumed profile.}
}
\value{
a matrix
}
\description{
Create the likelihood matrix of all possible response given a profile.
}
