% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_response_likelihood.R
\name{build_response_likelihood}
\alias{build_response_likelihood}
\title{Create the likelihood matrix for a matrix of probabilities of correct
response for non-masters and masters marginal only on that attribute.}
\usage{
build_response_likelihood(item_probs, response_patterns = NULL, prior = NULL)
}
\arguments{
\item{item_probs}{a matrix of probabilities for non-masters and masters.
Each row is an item.}

\item{response_patterns}{A binary matrix where each row is one response
pattern. If none is passed, then all possible patterns are generated}

\item{prior}{The prior for non-mastery and mastery ratio of the focus
attribute}
}
\value{
a matrix where the first columns contain the response patterns,
the likelihood of seeing that pattern given a certain level of mastery.
The probability of being a (non)master and the MLE (1 if estimated master,
0 otherwise).
}
\description{
Create the likelihood matrix for a matrix of probabilities of correct
response for non-masters and masters marginal only on that attribute.
}
