% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_matrix.R
\name{change_utils}
\alias{change_utils}
\alias{mat_is_complex}
\alias{mat_measures_attr}
\title{Utility functions for the \code{fun} and \code{at} parameters of \code{change_matrix}.}
\usage{
mat_is_complex(low, high = low)

mat_measures_attr(attr)
}
\arguments{
\item{low}{The lower range of complexity.}

\item{high}{The higher range of complexity. By default is the same as \code{low}
meaning it matches the vectors that have that exact complexity.}

\item{attr}{The number of the attr in the Q-matrix.}
}
\description{
Utility functions for the \code{fun} and \code{at} parameters of \code{change_matrix}.
}
