% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_monotonicity.R
\name{check_monotonicity}
\alias{check_monotonicity}
\title{Check that the monotonicity condition of the gdina model is fullfilled. This
is useful for when you don't want to force it, but you want to know if the
model was able to achieve it.}
\usage{
check_monotonicity(
  object,
  action = c("error", "warning", "message", "quiet"),
  tolerance = 1e-09
)
}
\arguments{
\item{object}{Either a a gdina object or a matrix of test_parameters}

\item{action}{what to do with the result. By default is a message.}

\item{tolerance}{tolerance for difference}
}
\value{
a boolean and raises a warning
}
\description{
Check that the monotonicity condition of the gdina model is fullfilled. This
is useful for when you don't want to force it, but you want to know if the
model was able to achieve it.
}
