% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_patterns.R
\name{create_patterns}
\alias{create_patterns}
\title{Create all possible permutations of the levels passed.}
\usage{
create_patterns(
  num_vars,
  include_filter = NULL,
  exclude_filter = NULL,
  levels = c(0, 1),
  column_prefix = "V"
)
}
\arguments{
\item{num_vars}{number of columns}

\item{include_filter}{vector of length \code{length(num_vars)}. It only includes rows that match. NA values work like wildcards.}

\item{exclude_filter}{vector of length \code{length(num_vars)}. It excludes rows that match. NA values work like wildcards.}

\item{levels}{Elements to be used in the combination. By default \code{c(0, 1)}, thus creating a binary matrix.}

\item{column_prefix}{The name of the output columns.}
}
\value{
a matrix of all possible permutations.
}
\description{
Create all possible permutations of the levels passed.
}
