% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_qmatrix.R
\name{create_qmatrix}
\alias{create_qmatrix}
\title{Create Q-Matrix}
\usage{
create_qmatrix(
  num_attr = 3L,
  items_per_type = 1,
  attributes.names = NULL,
  item.names = NULL
)
}
\arguments{
\item{num_attr}{attributes in the Q-matrix}

\item{items_per_type}{number or vector with the amount of items per unique combination}

\item{attributes.names}{vector of names for the attributes. Defaults to "Attr#"}

\item{item.names}{vector of names for the items. Defaults to "Item#"}
}
\value{
a Q-matrix
}
\description{
Create Q-Matrix
}
\examples{
# example code
# This creates a 4 attribute qmatrix with 3 items per attribute.
create_qmatrix(4, 3)

# This creates a 3 attribute qmatrix with 2 simple items per attribute,
# 1 double attribute item per combination, and 1 triple attribute item.
create_qmatrix(3, c(2, 1, 1))
}
