% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test.R
\name{create_test}
\alias{create_test}
\title{Create a matrix of items}
\usage{
create_test(..., test_type = "LCDM", num_attrs = NULL)
}
\arguments{
\item{...}{integer vectors with the probabilities}

\item{test_type}{to be passed to the item creator (all the same type).}

\item{num_attrs}{Needed to determine the matrix cols.}
}
\value{
a matrix of parameters
}
\description{
Create a matrix of items
}
\examples{
create_test(
  c(.2, .3, .5),
  c(.2, .4, .6),
  test_type="CRUM",
  num_attrs=2)
}
