% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_examinees.R, R/generate_family.R,
%   R/generate_responses.R
\name{generate_examinees}
\alias{generate_examinees}
\alias{generate_family}
\alias{generate_responses}
\title{Generate sample of examinees}
\usage{
generate_examinees(
  sample_size,
  total_attrs,
  base_rate = 0.5,
  attr_corr = 0,
  strict = TRUE,
  tolerance = NULL,
  attributes_names = NULL,
  responses_names = NULL
)

generate_responses(examinees, test, get_probs = FALSE, qmatrix = NULL)
}
\arguments{
\item{sample_size}{How many examinees to generate}

\item{total_attrs}{How many attributes are in the sample}

\item{base_rate}{Base rate per attribute (or one if it's the same for all)}

\item{attr_corr}{Correlation of attributes (or one if it's the same for all)}

\item{strict}{Boolean. If true, the function tries to generate a
binary matrix with the given correlation. If it is not possible, then it
raises and error. If false it is considered to be the correlation matrix of
the underlying normal distributions.}

\item{tolerance}{When strict_binary_corr=FALSE, the actual correlation values
can be far from the expected values. This gives a tolerance to avoid
unexpected results.}

\item{attributes_names}{vector of names for the attributes. Defaults to
"Attr#"}

\item{responses_names}{vector of names for the responses. Defaults to "ID#"}

\item{examinees}{Binary matrix of respondents}

\item{test}{matrix of items parameters}

\item{get_probs}{if you want the probability of correct response matrix.}

\item{qmatrix}{Optional. All test are checked to see if they conform to some
Q-matrix. If some specific qmatrix gets passed, then it is checked against
that one.}
}
\value{
A matrix of respondents and attributes.

a matrix of responses in the form 0/1 or the real probabilities.
}
\description{
The \verb{generate_*} functions are the ones that help start the simulation cycle
by generating the examinees and simulating their responses.
\itemize{
\item \code{\link[=generate_examinees]{generate_examinees()}} Create a matrix of examinees to use in simulations
\item \code{\link[=generate_responses]{generate_responses()}} Given a \link{test_parameters} matrix and a matrix of
examinees.
}
}
\keyword{internal}
