% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_responses.R
\name{generate_responses}
\alias{generate_responses}
\title{Generate responses to a test}
\usage{
generate_responses(qmat, respondents, items, get_probs = FALSE)
}
\arguments{
\item{qmat}{Q-matrix}

\item{respondents}{Binary matrix of respondents}

\item{items}{matrix of items-parameters}

\item{get_probs}{if you want the probability of correct response matrix.}
}
\value{
a matrix of responses in the form 0/1 or the real probabilities.
}
\description{
Given a Q-matrix and a sample of respondents this function returns a simulated
response matrix. Items parameter will automatically be mapped to the right attribute.
}
