% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_examinees.R, R/get_family.R,
%   R/get_test_parameters.R
\name{get_examinees}
\alias{get_examinees}
\alias{get_family}
\alias{get_test_parameters}
\title{get_* Functions}
\usage{
get_examinees(model, method = c("map", "mle"))

get_test_parameters(model, complete = TRUE, pretty_print = TRUE, digits = 2)
}
\arguments{
\item{model}{A fitted GDINA model.}

\item{method}{The method used in classification.}

\item{complete}{If the model only estimated some parameters this will not
show in the table. So, for example, the ACDM model will only return main
effects. If true this returns the complete table and fills the corresponding
parameters with 0. This also ensures a consistent order of the parameters.}

\item{pretty_print}{boolean. If true it prints a pretty table. It returns a
data frame silently.}

\item{digits}{rounding digits when printing. The return is not approximated.}
}
\value{
A matrix with the MAP or MLE classifications of examinees.

a \link{test_parameters} object from the parameters estimated by the
model, which is just an extension of the base \link[base:matrix]{base::matrix} class.
}
\description{
The \verb{get_*} functions take as a first argument a fitted GDINA model and
return a matrix to simplify piping operations.
\itemize{
\item \code{\link[=get_test_parameters]{get_test_parameters()}} returns a \link{test_parameters} object from the
parameters estimated by the model.
\item \code{\link[=get_examinees]{get_examinees()}} returns a numeric matrix of the students classification
by attribute.
}
}
\concept{examinees functions}
\concept{test parameters functions}
\keyword{internal}
