% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_probabilities_by_profile.R
\name{get_probabilities_by_profile}
\alias{get_probabilities_by_profile}
\title{Returns a matrix of profiles and the probability of a correct response for all items.}
\usage{
get_probabilities_by_profile(model)
}
\arguments{
\item{model}{A GDINA object}
}
\value{
a matrix of profiles and the probability of a correct response for all items.
}
\description{
Returns a matrix of profiles and the probability of a correct response for all items.
}
