% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saltr-package.R
\docType{package}
\name{saltr-package}
\alias{saltr}
\alias{saltr-package}
\title{saltr: Simulation And Leaplab Tools in R}
\description{
Simulation And Leaplab Tools in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://beltranpantoja.github.io/saltR/}
  \item \url{https://github.com/beltranpantoja/saltR}
}

}
\author{
\strong{Maintainer}: Beltrán Pantoja \email{bpp09644@uga.edu}

}
\keyword{internal}
