% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_parameters.R
\name{test_parameters}
\alias{test_parameters}
\alias{as_test_parameters}
\alias{pretty_print}
\title{Test parameters}
\usage{
as_test_parameters(x)

pretty_print(test_parameters, digits = 3, prefix = NULL)
}
\arguments{
\item{x}{A matrix or a numeric data.frame.}

\item{test_parameters}{matrix to print}

\item{digits}{number of rounding digits}

\item{prefix}{prefix for the columns. By default, uses the Greek letter
lambda in interactive sessions.}
}
\value{
the passed test_parameters with no change
}
\description{
A \code{test_parameters} object is a matrix used to store item/test parameters
with enhanced printing and validation.
}
\concept{test parameter objects}
